requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
